Removed unnecessary initializations when setting Tx or Rx mode.
The same initialization of settings and pins is done during Init(). The change causes
an significant speed improvement when switching between Tx and Rx modes rapidly.
According to figure 13 (state diagram) in C1101 manual, it is allowed to go
directly to Tx and Rx from either Idle, Tx or Rx mode.